requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
databricks-sdk
pydantic
python-multipart
jinja2